[backend] refacto completeContextDataForEntity (#5611)
